M4-A corrections: fix cursor key modifiers, test buffer aliasing, mouse coordinates, docs
Issues corrected:

1. HIGH: Cursor key modifier encoding was incomplete and incorrect.
   - Now handles Shift modifier (previously only Ctrl/Alt checked)
   - Fixed CSI parameter calculation: use (1 + mod) not ('0' + mod)
   - Supports all modifier combinations per standard CSI format

2. HIGH: Determinism/reset tests had false-positive buffer aliasing.
   - Tests now copy byte slices to separate buffers before comparing
   - Can now properly detect regressions in encoding behavior
   - Prevents aliasing to reused internal encode_buf

3. MEDIUM: Mouse coordinate type mismatch with contract.
   - Changed MouseEvent.row from usize to i32 (supports negative history indices)
   - Changed MouseEvent.col from usize to u16 (viewport range)
   - Matches selection coordinate model (signed row, unsigned col)
   - Added test validating history row support

4. MEDIUM: Tracking documents not updated to reflect M4-A completion.
   - Updated MILESTONE.md to mark M4-A1/A2/A3 as complete
   - Updated ACTIVE_QUEUE.md with M4-A closeout summary
   - Documented technical corrections made

Contract updates:
- INPUT_CONTROL.md: Clarified cursor key modifier encoding (CSI 1;(1+mod)letter)
- MODEL_API.md: Documented mouse coordinate types and semantics

All validation passed: zig build, zig build test, shim grep clean.
391 tests pass.

+++ src/model/types.zig
@@ -110,8 +110,8 @@ pub const MouseEventKind = enum(u8) {
 pub const MouseEvent = struct {
     kind: MouseEventKind,
     button: MouseButton,
-    row: usize,
-    col: usize,
+    row: i32,
+    col: u16,
     pixel_x: ?u32 = null,
     pixel_y: ?u32 = null,
     mod: Modifier,
